feat: Add script to fetch and print all states from the database
The code changes include adding a new script `7-model_state_fetch_all.py` that fetches all the states from the database and prints their IDs and names. This script will be useful for retrieving state information from the database.

Based on the recent user commits and repository commits, it seems that there have been some refactorings and fixes in the `model_state.py` file. However, these changes are not directly related to the new script.

Please review and consider this commit message.